moved pkg lib/colibri to lib/colibri/reservation
