+++ go/lib/colibri/reservation/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["types.go"],
-    importpath = "github.com/scionproto/scion/go/lib/colibri",
+    importpath = "github.com/scionproto/scion/go/lib/colibri/reservation",
     visibility = ["//visibility:public"],
     deps = [
         "//go/lib/addr:go_default_library",
@@ -16,8 +16,5 @@ go_test(
     name = "go_default_test",
     srcs = ["types_test.go"],
     embed = [":go_default_library"],
-    deps = [
-        "//go/lib/common:go_default_library",
-        "//go/lib/xtest:go_default_library",
-    ],
+    deps = ["//go/lib/xtest:go_default_library"],
 )
