+++ go/lib/colibri/reservation/types.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package colibri
+package reservation
 
 import (
 	"github.com/scionproto/scion/go/lib/addr"
